+++ fixture-data/src/nextest_tests.rs
@@ -22,7 +22,8 @@ pub static EXPECTED_TEST_SUITES: Lazy<BTreeMap<RustBinaryId, TestSuiteFixture>> 
             "basic",
             BuildPlatform::Target,
             vec![
-                TestCaseFixture::new("test_cargo_env_vars", TestCaseFixtureStatus::Pass),
+                TestCaseFixture::new("test_cargo_env_vars", TestCaseFixtureStatus::Pass)
+                    .with_property(TestCaseFixtureProperty::NotInDefaultSetUnix),
                 TestCaseFixture::new("test_cwd", TestCaseFixtureStatus::Pass)
                     .with_property(TestCaseFixtureProperty::NeedsSameCwd),
                 TestCaseFixture::new("test_execute_bin", TestCaseFixtureStatus::Pass),